+++ test/concrete/arb/GenericPoolOrderBookV4ArbOrderTaker.sender.t.sol
@@ -31,8 +31,8 @@ contract GenericPoolOrderBookV5ArbOrderTakerSenderTest is GenericPoolOrderBookV5
             iOrderBook,
             TakeOrdersConfigV4(
                 LibDecimalFloat.packLossless(0, 0),
-                LibDecimalFloat.packLossless(type(int256).max, 0),
-                LibDecimalFloat.packLossless(type(int256).max, 0),
+                LibDecimalFloat.packLossless(type(int224).max, 0),
+                LibDecimalFloat.packLossless(type(int224).max, 0),
                 orders,
                 abi.encode(iRefundoor, iRefundoor, "")
             ),
