+++ test/concrete/ob/OrderBook.takeOrder.handleIO.revert.t.sol
@@ -52,10 +52,10 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
             vm.mockCall(outputToken, abi.encodeWithSelector(IERC20.transfer.selector, address(this)), abi.encode(true));
             vm.mockCall(inputToken, "", abi.encode(true));
         }
-        iOrderbook.deposit3(outputToken, vaultId, LibDecimalFloat.packLossless(type(int256).max, -18), new TaskV2[](0));
+        iOrderbook.deposit3(outputToken, vaultId, LibDecimalFloat.packLossless(type(int224).max, -18), new TaskV2[](0));
         assertTrue(
             iOrderbook.vaultBalance2(address(this), outputToken, vaultId).eq(
-                LibDecimalFloat.packLossless(type(int256).max, -18)
+                LibDecimalFloat.packLossless(type(int224).max, -18)
             )
         );
 
@@ -75,7 +75,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
             orders[i] = TakeOrderConfigV4(order, 0, 0, new SignedContextV1[](0));
         }
         TakeOrdersConfigV4 memory takeOrdersConfig = TakeOrdersConfigV4(
-            LibDecimalFloat.packLossless(0, 0), maxInput, LibDecimalFloat.packLossless(type(int256).max, 0), orders, ""
+            LibDecimalFloat.packLossless(0, 0), maxInput, LibDecimalFloat.packLossless(type(int224).max, 0), orders, ""
         );
 
         if (err.length > 0) {
@@ -91,21 +91,21 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
     function testTakeOrderHandleIO00() external {
         bytes[] memory configs = new bytes[](1);
         configs[0] = "_ _:max-value() 1;:ensure(0 \"err\");";
-        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO1() external {
         bytes[] memory configs = new bytes[](2);
         configs[0] = "_ _:1 1;:ensure(0 \"err\");";
         configs[1] = "_ _:1 1;:;";
-        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO2() external {
         bytes[] memory configs = new bytes[](2);
         configs[0] = "_ _:1 1;:;";
         configs[1] = "_ _:1 1;:ensure(0 \"err\");";
-        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO3() external {
@@ -113,7 +113,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         configs[0] = "_ _:1 1;:;";
         configs[1] = "_ _:1 1;:ensure(0 \"err\");";
         configs[2] = "_ _:1 1;:;";
-        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO4() external {
@@ -121,7 +121,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         configs[0] = "_ _:1 1;:;";
         configs[1] = "_ _:1 1;:ensure(0 \"err 1\");";
         configs[2] = "_ _:1 1;:ensure(0 \"err 2\");";
-        checkTakeOrderHandleIO(configs, "err 1", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err 1", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO5() external {
@@ -129,7 +129,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         configs[0] = "_ _:1 1;:;";
         configs[1] = "_ _:1 1;:ensure(0 \"err 2\");";
         configs[2] = "_ _:1 1;:ensure(0 \"err 1\");";
-        checkTakeOrderHandleIO(configs, "err 2", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err 2", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     function testTakeOrderHandleIO6() external {
@@ -137,12 +137,12 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         configs[0] = "_ _:1 1;:ensure(0 \"err 2\");";
         configs[1] = "_ _:1 1;:;";
         configs[2] = "_ _:1 1;:ensure(0 \"err 1\");";
-        checkTakeOrderHandleIO(configs, "err 2", LibDecimalFloat.packLossless(type(int256).max, 0));
+        checkTakeOrderHandleIO(configs, "err 2", LibDecimalFloat.packLossless(type(int224).max, 0));
     }
 
     /// forge-config: default.fuzz.runs = 100
     function testTakeOrderHandleIO7(uint256 toClear18) external {
-        toClear18 = bound(toClear18, 3e18 + 1, uint256(type(int256).max));
+        toClear18 = bound(toClear18, 3e18 + 1, uint256(int256(type(int224).max)));
         Float toClear = LibDecimalFloat.fromFixedDecimalLosslessPacked(toClear18, 18);
         bytes[] memory configs = new bytes[](4);
         configs[0] = "_ _:1 1;:set(0 1);";
@@ -154,7 +154,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
 
     /// forge-config: default.fuzz.runs = 100
     function testTakeOrderHandleIO8(uint256 toClear18) external {
-        toClear18 = bound(toClear18, 4e18 + 1, uint256(type(int256).max));
+        toClear18 = bound(toClear18, 4e18 + 1, uint256(int256(type(int224).max)));
         Float toClear = LibDecimalFloat.fromFixedDecimalLosslessPacked(toClear18, 18);
         bytes[] memory configs = new bytes[](5);
         configs[0] = "_ _:1 1;:;";
@@ -202,7 +202,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         checkTakeOrderHandleIO(
             configs,
             abi.encodeWithSelector(SourceIndexOutOfBounds.selector, 1, hex"010000020200020110000001100000"),
-            LibDecimalFloat.packLossless(type(int256).max, 0)
+            LibDecimalFloat.packLossless(type(int224).max, 0)
         );
     }
 
@@ -215,7 +215,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         checkTakeOrderHandleIO(
             configs,
             abi.encodeWithSelector(SourceIndexOutOfBounds.selector, 1, hex"010000020200020110000001100000"),
-            LibDecimalFloat.packLossless(type(int256).max, 0)
+            LibDecimalFloat.packLossless(type(int224).max, 0)
         );
     }
 
@@ -228,7 +228,7 @@ contract OrderBookTakeOrderHandleIORevertTest is OrderBookExternalRealTest {
         checkTakeOrderHandleIO(
             configs,
             abi.encodeWithSelector(SourceIndexOutOfBounds.selector, 1, hex"010000020200020110000001100000"),
-            LibDecimalFloat.packLossless(type(int256).max, 0)
+            LibDecimalFloat.packLossless(type(int224).max, 0)
         );
     }
 }
